requirements.txt
Flask==2.0.1
supabase==2.3.1
python-dotenv==0.19.0
qrcode==7.3.1
Pillow==8.3.1